Add support for multiple email recipients
Updated the EmailClient and IEmailClient to support multiple recipients when sending an email. A new enumeration, RecipientType was created for recipient type identification (To, Cc, Bcc) and a helper function, AddRecipients, was added to handle recipient addition to the MimeMessage. This change enhances the email sending function by allowing emails to be sent to multiple recipients of different types.